Refactor: Modularize Gerchberg-Saxton algorithm and perturbations
